Move load_clip module one level up
