+++ clip_retrieval/clip_back.py
@@ -854,7 +854,7 @@ def load_mclip(clip_model):
 def load_clip_index(clip_options):
     """load the clip index"""
     import torch  # pylint: disable=import-outside-toplevel
-    from .clip_inference.load_clip import load_clip  # pylint: disable=import-outside-toplevel
+    from clip_retrieval.load_clip import load_clip  # pylint: disable=import-outside-toplevel
 
     device = "cuda" if torch.cuda.is_available() else "cpu"
     model, preprocess = load_clip(clip_options.clip_model, use_jit=clip_options.use_jit, device=device)
