+++ clip_retrieval/clip_inference/worker.py
@@ -14,9 +14,9 @@ from braceexpand import braceexpand
 from clip_retrieval.clip_inference.runner import Runner
 from clip_retrieval.clip_inference.mapper import ClipMapper
 from clip_retrieval.clip_inference.writer import NumpyWriter
-from clip_retrieval.clip_inference.load_clip import load_clip
 from clip_retrieval.clip_inference.logger import LoggerWriter
 from clip_retrieval.clip_inference.reader import FilesReader, WebdatasetReader
+from clip_retrieval.load_clip import load_clip
 
 
 def worker(
