+++ tests/test_clip_inference/test_runner.py
@@ -3,7 +3,7 @@ from clip_retrieval.clip_inference.runner import Runner
 from clip_retrieval.clip_inference.reader import FilesReader
 from clip_retrieval.clip_inference.mapper import ClipMapper
 from clip_retrieval.clip_inference.writer import NumpyWriter
-from clip_retrieval.clip_inference.load_clip import load_clip
+from clip_retrieval.load_clip import load_clip
 import os
 import numpy as np
 import tempfile
